+++ components/dynamic-font-loader.tsx
@@ -2,24 +2,25 @@
 
 import { useMounted } from "@/hooks/use-mounted";
 import { useEditorStore } from "@/store/editor-store";
-import { extractFontFamily, getDefaultWeights, loadGoogleFont } from "@/utils/fonts";
+import { extractFontFamily, getDefaultWeights } from "@/utils/fonts";
+import { loadGoogleFont } from "@/utils/fonts/google-fonts";
 import { useEffect, useMemo } from "react";
 
 export function DynamicFontLoader() {
   const { themeState } = useEditorStore();
   const isMounted = useMounted();
 
+  const fontSans = themeState.styles.light["font-sans"];
+  const fontSerif = themeState.styles.light["font-serif"];
+  const fontMono = themeState.styles.light["font-mono"];
+
   const currentFonts = useMemo(() => {
     return {
-      sans: themeState.styles.light["font-sans"],
-      serif: themeState.styles.light["font-serif"],
-      mono: themeState.styles.light["font-mono"],
+      sans: fontSans,
+      serif: fontSerif,
+      mono: fontMono,
     } as const;
-  }, [
-    themeState.styles.light["font-sans"],
-    themeState.styles.light["font-serif"],
-    themeState.styles.light["font-mono"],
-  ]);
+  }, [fontSans, fontSerif, fontMono]);
 
   useEffect(() => {
     if (!isMounted) return;
