+++ components/theme-script.tsx
@@ -5,7 +5,7 @@ import { defaultDarkThemeStyles, defaultLightThemeStyles } from "@/config/theme"
 export function ThemeScript() {
   const scriptContent = `
     // ----- FONT LOADING UTILITIES -----
-    const DEFAULT_FONT_WEIGHTS = ["400", "500", "600", "700"];
+    const DEFAULT_FONT_WEIGHTS = ["400"];
 
     function extractFontFamily(fontFamilyValue) {
       if (!fontFamilyValue) return null;
